src/config/settings.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    API_KEY: str = os.getenv("API_KEY")
    TWILIO_ACCOUNT_SID: str = os.getenv("TWILIO_ACCOUNT_SID")
    TWILIO_AUTH_TOKEN: str = os.getenv("TWILIO_AUTH_TOKEN")
    GOOGLE_CLOUD_STORAGE_BUCKET: str = os.getenv("GOOGLE_CLOUD_STORAGE_BUCKET")
    GCP_PROJECT_ID: str = os.getenv("GCP_PROJECT_ID")
    OPENAI_API_KEY: str = os.getenv("OPENAI_API_KEY")

settings = Settings()