requirements.txt
fastapi
uvicorn
pydantic
twilio
google-cloud-storage
openai
python-dotenv
pytest